Allow value prop to be set to state of parent
Previously, the value of the input could not be updated through the
state of its parent. It allowed only a single edit operation per
debounce because the parent's state would keep updating the
input's value to the pre-debounce value inbetween key-strokes.
By now comparing the received props to the input's state, it will
only update the input's state if the parent's state has changed.