Fix bug with sticky onChange props
I noticed some weird behavior when I had the onChange handlers changing dynamically, where it would only invoke the first onChange function.

To fix, I just made the `doNotify()` function that calls `this.props.onChange`. This way old references to the old debounced onChange handler don't hang around and it always calls the current onChange.

The downside to this approach is if the onChange handler changes between an event and the debounced call invocation, then it will invoke the new handler, not the one that was set when the event first took place. If you feel this is a bad tradeoff, feel free to close this PR.