Set isDebouncing flag before first render on update
When migrating from componentWillReceiveProps to componentDidUpdate
the initial state update on the first keypress was being processed
before the isDebouncing flag was set.

In order to combat hitting a check that the isDebouncing flag was
intended to avoid, an additional check was added that causes
undesirable rendering behavior in certain cases where the internal
state can get out of sync with the external values, especially
during external validation that might want to update the component's
value prop.

To mitigate, we need to set the isDebouncing flag before we set the
initial state.